[IDP-2202] modified barcodeAuth method and refactor common class
