+++ src/main/java/it/gov/pagopa/common/utils/CommonUtilities.java
@@ -1,18 +1,13 @@
 package it.gov.pagopa.common.utils;
 
-import it.gov.pagopa.payment.dto.Reward;
-import it.gov.pagopa.payment.model.counters.RewardCounters;
-import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
-import java.util.Map;
 
 public class CommonUtilities {
-  private static final Pair<Boolean, Long> PAIR_NULL = Pair.of(null,null);
   private CommonUtilities() {}
 
   /** To convert cents into euro */
@@ -33,28 +28,7 @@ public class CommonUtilities {
     return pageable;
   }
 
-  public static Long calculateResidualBudget(Map<String, Reward> rewards) {
-    Long residualBudget = null;
-    Reward reward = rewards.values().stream().findFirst().orElse(null);
-    RewardCounters rewardCounters = reward != null ? reward.getCounters() : null;
-    if (reward != null && rewardCounters != null) {
-      residualBudget = euroToCents(rewardCounters.getInitiativeBudget().subtract(rewardCounters.getTotalReward()));
-    }
-    return residualBudget;
-  }
-
   public static Long minutesToSeconds(Integer minutes){
     return minutes == null ? null : (long)minutes*60;
-  }
-
-  /** idpaycode or qrcode = reward, barcode = residualBudget */
-  public static Pair<Boolean, Long> getSplitPaymentAndResidualAmountCents(Long amountCents, Long rewardOrResidualBudget) {
-    if (rewardOrResidualBudget != null) {
-      long residualAmountCents = amountCents - rewardOrResidualBudget;
-      if (residualAmountCents >= 0L) {
-        return Pair.of(residualAmountCents > 0L, residualAmountCents);
-      }
-    }
-    return PAIR_NULL;
   }
 }
