+++ src/main/java/it/gov/pagopa/payment/dto/mapper/TransactionInProgress2SyncTrxStatusMapper.java
@@ -1,9 +1,9 @@
 package it.gov.pagopa.payment.dto.mapper;
 
-import it.gov.pagopa.common.utils.CommonUtilities;
 import it.gov.pagopa.payment.dto.qrcode.SyncTrxStatusDTO;
 import it.gov.pagopa.payment.enums.SyncTrxStatus;
 import it.gov.pagopa.payment.model.TransactionInProgress;
+import it.gov.pagopa.payment.utils.CommonPaymentUtilities;
 import it.gov.pagopa.payment.utils.RewardConstants;
 import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.stereotype.Service;
@@ -21,7 +21,7 @@ public class TransactionInProgress2SyncTrxStatusMapper {
 
     public SyncTrxStatusDTO transactionInProgressMapper(TransactionInProgress transaction){
 
-        Pair<Boolean, Long> splitAndResidualAmountCents = CommonUtilities.getSplitPaymentAndResidualAmountCents(transaction.getAmountCents(), transaction.getReward());
+        Pair<Boolean, Long> splitAndResidualAmountCents = CommonPaymentUtilities.getSplitPaymentAndResidualAmountCents(transaction.getAmountCents(), transaction.getReward());
 
         SyncTrxStatusDTO response = SyncTrxStatusDTO.builder()
                 .id(transaction.getId())
