+++ src/main/java/it/gov/pagopa/payment/service/payment/barcode/BarCodeCreationServiceImpl.java
@@ -76,11 +76,12 @@ public class BarCodeCreationServiceImpl extends CommonCreationServiceImpl implem
 
             TransactionInProgress trx =
                     transactionBarCodeCreationRequest2TransactionInProgressMapper.apply(
-                            trxBarCodeCreationRequest, channel, userId, initiative != null ? initiative.getInitiativeName() : null, residualBudgetCents);
+                            trxBarCodeCreationRequest, channel, userId, initiative != null ? initiative.getInitiativeName() : null);
             generateTrxCodeAndSave(trx);
 
             logCreatedTransaction(trx.getInitiativeId(), trx.getId(), trx.getTrxCode(), userId);
 
+            trx.setAmountCents(residualBudgetCents);
             return transactionBarCodeInProgress2TransactionResponseMapper.apply(trx);
 
         } catch (RuntimeException e) {
